requirements.txt
python-telegram-bot[job-queue]==20.7
beautifulsoup4
lxml
requests
python-dotenv
schedule
nest_asyncio